+++ src/FlightTrend.Serializers.Test/ReturnFlightArchiveSerializerTest.cs
@@ -13,7 +13,7 @@ namespace FlightTrend.Serializers.Test
                 new InstantSerializer(),
                 new LocalDateSerializer(),
                 new LocalTimeSerializer(),
-                new DecimalSerializer());
+                new FloatSerializer());
 
         [TestMethod]
         public void SerializeAndDeserializeShouldReturnTheSame()
